E_commerce/requirements.txt
django
psycopg2-binary
djangorestframework
django-jazzmin
python-dotenv
requests
pandas
django-debug-toolbar
elasticsearch==7.17.0
django-elasticsearch-dsl==7.2.2
celery
django-celery-beat
redis